requirements.txt
altair==5.1.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
dask==2023.9.2
faiss-gpu==1.7.2
filelock==3.9.0
fsspec==2023.4.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.2.1
networkx==3.0rc1
numpy==1.26.0
packaging==23.1
pandas==2.1.1
partd==1.4.1
Pillow==9.5.0
protobuf==4.24.3
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytorch-triton==2.1.0+6e4932cda8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
six==1.16.0
smmap==5.0.1
streamlit==1.27.0
sympy==1.11.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==2.2.0.dev20230927+cu121
torchaudio==2.2.0.dev20230927+cu121
torchvision==0.17.0.dev20230927+cu121
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.17.0
